redid the way that the investions worked
There was a problem with the previous way. now I am marking each of the
Inverted images with a uninvested image (this time only in the header,
the data is the same) this uninverted image can be used to figure out
the translation to align centers of mass. in the registration Phase, if
it is detected that an uninverted image exists, this translate transform
is calculated, and  is concatenated with a flip transform. this
concatenated transform is used as the initial transform for the affine
registration.